Add Student and Mentor Controllers, and Student registration and welcome views
In this commit, two new controllers for Student and Mentor as well as views for student registration and welcome page have been created. The JavaFaker library has been added as a dependency in pom.xml for data generation. Additionally, a new Student model and a DataGenerator class to generate student data are also implemented.