Add new LabController classes and update cart-list template
This commit introduces two new controller classes, LabController01 and LabController02. LabController01 handles the 'lab' and 'profile' endpoints, and it also provided profile information and lab listings. LabController02 is dedicated to the 'cart-list' endpoint, which includes the functionality to display all cart items. A minor change was also made to the cart-list.html to improve code readability regarding size checking.