Implement Thymeleaf dynamic elements in templates
Thymeleaf dynamic elements are added to the HTML templates in the SpringBoot-Thymeleaf-Lab, displaying data from the product list. Conditionally render sections in the 'list.html' file based on the size of the productList. Adding URL paths to the 'add to cart' and 'see cart' buttons in 'list.html'. Also, method to return the list view is added in ProductController.