Add pizza ordering functionality and related classes
This commit introduces a myriad of new classes related to a pizza ordering feature in a Spring Boot application. It lays the groundwork for the application by defining various enums for the pizza attributes like Cheese, Sauce, and Topping. It also sets up a DataGenerator for pre-defining some data for these attributes. In addition, it introduces Controllers for designing and ordering the pizza and implements related models such as the Pizza and PizzaOrder.