Refactor and fix bugs in Pizza Cloud application
Implemented necessary changes in the code to fix the bugs. Refactored the PizzaRepository to handle exceptions when a pizza is not found. Also, reorganized the HTML templates, especially the form actions to accurately capture user inputs. In addition, various adjustments for clean code standards were made in the enums and models. Some changes have been made in the controllers as well, to correct the way the request handling was originally set up.