Add Merchant entity and update Payment and DataGenerator classes
A Merchant entity has been added to create a relationship between Merchants and Payments. It has been annotated as a JPA entity and all necessary fields have been incorporated. The Payment class has been updated to include a Many-to-One relationship with the Merchant entity. In addition, the DataGenerator class has been updated to include debugging print statements for testing.