Add initial data to the database
A comprehensive set of initial data has been added to the 'departments', 'regions', and 'employees' tables in the database. This data encompasses various departments, regions across multiple countries, and a detailed list of employees. This initial data will be crucial for further development and testing of features that interact with these tables.